backend/requirements.txt
asgiref==3.11.1
Django==6.0.5
django-cors-headers==4.9.0
djangorestframework==3.17.1
pyswip==0.3.3
sqlparse==0.5.5
tzdata==2026.2
requests==2.31.0
python-dotenv==1.0.0
